Reverted CourseCard move
